requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0

# D-Wave dependencies
dwave-ocean-sdk>=6.0.0
dwave-pytorch-plugin>=0.1.0
dimod>=0.12.0
